12_Challenge_Financial_Agent/pyproject.toml
[project]
name = "12-challenge-finantial-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "gradio>=5.49.1",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "pandas>=2.3.3",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.1.2",
    "tabulate>=0.9.0",
]
